🚚 Rename package paths of Week06.EIZI
