+++ 공통주차별문제/Week06/공통문제/EIZI/boj1389케빈베이컨의6단계법칙.java
@@ -1,4 +1,4 @@
-package BAEKJOON;
+package 공통주차별문제.Week06.공통문제.EIZI;
 
 import java.io.BufferedReader;
 import java.io.IOException;
