+++ 공통주차별문제/Week06/공통문제/EIZI/boj15486퇴사2.java
@@ -1,4 +1,4 @@
-package BAEKJOON;
+package 공통주차별문제.Week06.공통문제.EIZI;
 
 import java.io.BufferedReader;
 import java.io.IOException;
