+++ 공통주차별문제/Week06/공통문제/EIZI/boj17615볼모으기.java
@@ -1,8 +1,10 @@
+package 공통주차별문제.Week06.공통문제.EIZI;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class Main {
+public class boj17615볼모으기 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
